chore: move navigation to its own file
