+++ lib/App.js
@@ -3,110 +3,11 @@ import {
   createAppContainer,
   createStackNavigator,
   createSwitchNavigator,
-  createBottomTabNavigator,
 } from 'react-navigation'
-import Home from './screens/Home/Home'
-import Settings from './screens/Settings/Settings'
 import Wizard from './screens/Account/Account'
-import Auth from './screens/Auth/Auth'
 import AuthLoading from './screens/AuthLoading/AuthLoading'
-import styled, { ThemeProvider, theme as myDataTheme } from './theme'
-import { NavbarIcon } from './components/elements/Icon/Icon'
-import { isIphoneX } from './utils/is-iphone-x'
-import { View, TouchableOpacity } from 'react-native'
-
-const NavItemQRBackground = styled(View)`
-  background-color: ${({ theme }) => theme.colors.quiet};
-  border-radius: 65;
-  border: 4px solid ${({ theme }) => theme.colors.quiet};
-  margin-bottom: 15px;
-`
-
-const NavItemQR = styled(TouchableOpacity)`
-  height: 100px;
-  width: 100px;
-  align-items: center;
-  justify-content: center;
-`
-
-const NavItem = styled(TouchableOpacity)`
-  margin-bottom: 15px;
-`
-
-const TabBarContainer = styled(View)`
-  height: 125px;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-  justify-content: space-evenly;
-  background: ${({ theme }) => theme.colors.quiet};
-  position: relative;
-`
-
-const TabBarInnerBlock = styled(View)`
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  height: 55;
-  background: ${({ theme }) => theme.colors.white};
-`
-
-const iphoneXTabBarStyles = isIphoneX() ? { height: 65 } : {}
-const iphoneXIconStyles = isIphoneX() ? { marginBottom: 22 } : {}
-const iphoneXQRIconStyles = isIphoneX() ? { marginBottom: 25 } : {}
-
-const getTabBarIcon = (routeName, isActive) => (
-  <NavbarIcon name={routeName.toLowerCase()} active={isActive} />
-)
-
-const getTabBarItem = (route, routeIndex, currentNavIndex, onTabPress) => {
-  const isRouteActive = routeIndex === currentNavIndex
-
-  return route.key === 'QR-kod' ? (
-    <NavItemQRBackground key={route.key} style={{ ...iphoneXQRIconStyles }}>
-      <NavItemQR onPress={() => onTabPress({ route })}>
-        {getTabBarIcon(route.key, isRouteActive)}
-      </NavItemQR>
-    </NavItemQRBackground>
-  ) : (
-    <NavItem
-      onPress={() => onTabPress({ route })}
-      style={{ ...iphoneXIconStyles }}
-      key={route.key}
-    >
-      {getTabBarIcon(route.key, isRouteActive)}
-    </NavItem>
-  )
-}
-
-const TabBar = ({ navigation, onTabPress }) => (
-  <TabBarContainer>
-    <TabBarInnerBlock style={{ ...iphoneXTabBarStyles }} />
-    {navigation.state.routes.map((route, routeIndex) =>
-      getTabBarItem(route, routeIndex, navigation.state.index, onTabPress)
-    )}
-  </TabBarContainer>
-)
-
-const TabBarWithDisplayName = props => <TabBar {...props} />
-
-const AppStack = createBottomTabNavigator(
-  {
-    Hem: { screen: Home },
-    'QR-kod': {
-      screen: Auth,
-      navigationOptions: () => ({
-        // Hides on camera view
-        tabBarVisible: false,
-      }),
-    },
-    Konto: { screen: Settings },
-  },
-  {
-    tabBarComponent: props => TabBarWithDisplayName(props),
-  }
-)
+import BottomTabNavigator from './Navigation'
+import { ThemeProvider, theme as myDataTheme } from './theme'
 
 const RegisterStack = createStackNavigator({
   Register: { screen: Wizard, navigationOptions: () => ({ header: null }) },
@@ -116,7 +17,7 @@ const AppSwitchContainer = createAppContainer(
   createSwitchNavigator(
     {
       AuthLoading,
-      App: AppStack,
+      App: BottomTabNavigator,
       Register: RegisterStack,
     },
     {
