+++ lib/components/elements/Icon/Icon.js
@@ -21,10 +21,6 @@ const HomeIcon = ({ active }) => (
 )
 
 const ConsentIcon = () => (
-  // <Image
-  //   style={{ width: 125, height: 125 }}
-  //   source={require('./img/qricon.png')}
-  // />
   <StyledQRIcon>
     <Icon
       name={'qr'}
